move splatoon 2 model classes to an own package
