+++ src/main/java/tv/strohi/twitch/strohkoenigbot/data/model/splatoon2/splatoondata/SplatoonAbilityMatch.java
@@ -1,31 +1,26 @@
-package tv.strohi.twitch.strohkoenigbot.data.model.splatoondata;
+package tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata.enums.SplatoonGearType;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import java.time.Instant;
 
 @Entity
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class SplatoonWeaponRequestRanking {
+public class SplatoonAbilityMatch {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private long id;
 
-	private String twitchId;
-
-	private String twitchName;
-
-	private Integer winStreak;
-
-	private Long weaponId;
-
-	private Instant challengedAt;
+	private Long abilityId;
+	private Long matchId;
+	private SplatoonGearType kind;
+	private Integer gearPosition;
 }
