+++ src/main/java/tv/strohi/twitch/strohkoenigbot/data/model/splatoon2/splatoondata/SplatoonClip.java
@@ -1,44 +1,38 @@
-package tv.strohi.twitch.strohkoenigbot.data.model.splatoondata;
+package tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import tv.strohi.twitch.strohkoenigbot.data.model.splatoondata.enums.SplatoonMode;
-import tv.strohi.twitch.strohkoenigbot.data.model.splatoondata.enums.SplatoonRule;
 
 import javax.persistence.*;
 import java.time.Instant;
 
 @Entity
+@Cacheable(false)
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class SplatoonRotation {
+public class SplatoonClip {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private long id;
-
-	private Long splatoonApiId;
 
 	private Long startTime;
 
 	private Long endTime;
 
-	private Long stageAId;
+	private Boolean isGoodPlay;
 
-	private Long stageBId;
+	private String clipUrl;
 
-	private SplatoonMode mode;
+	private String description;
 
-	private SplatoonRule rule;
+	private Long matchId;
 
-	@JsonIgnore
 	public Instant getStartTimeAsInstant() {
 		return Instant.ofEpochSecond(startTime);
 	}
 
-	@JsonIgnore
 	public Instant getEndTimeAsInstant() {
 		return Instant.ofEpochSecond(endTime);
 	}
