+++ src/main/java/tv/strohi/twitch/strohkoenigbot/data/model/splatoon2/splatoondata/SplatoonGear.java
@@ -1,8 +1,9 @@
-package tv.strohi.twitch.strohkoenigbot.data.model.splatoondata;
+package tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata.enums.SplatoonGearType;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -13,24 +14,16 @@ import javax.persistence.Id;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class SplatoonWeapon {
+public class SplatoonGear {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private long id;
 
 	private String splatoonApiId;
+
+	private SplatoonGearType kind;
+
 	private String name;
+
 	private String image;
-
-	private String subSplatoonApiId;
-	private String subName;
-	private String subImage;
-
-	private String specialSplatoonApiId;
-	private String specialName;
-	private String specialImage;
-
-	private Long turf;
-	private Integer wins;
-	private Integer defeats;
 }
