+++ src/main/java/tv/strohi/twitch/strohkoenigbot/data/model/splatoon2/splatoondata/SplatoonWeaponRequestRanking.java
@@ -1,26 +1,31 @@
-package tv.strohi.twitch.strohkoenigbot.data.model.splatoondata;
+package tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import tv.strohi.twitch.strohkoenigbot.data.model.splatoondata.enums.SplatoonGearType;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.time.Instant;
 
 @Entity
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class SplatoonAbilityMatch {
+public class SplatoonWeaponRequestRanking {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private long id;
 
-	private Long abilityId;
-	private Long matchId;
-	private SplatoonGearType kind;
-	private Integer gearPosition;
+	private String twitchId;
+
+	private String twitchName;
+
+	private Integer winStreak;
+
+	private Long weaponId;
+
+	private Instant challengedAt;
 }
