+++ src/main/java/tv/strohi/twitch/strohkoenigbot/data/repository/splatoondata/SplatoonClipRepository.java
@@ -2,7 +2,7 @@ package tv.strohi.twitch.strohkoenigbot.data.repository.splatoondata;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import tv.strohi.twitch.strohkoenigbot.data.model.splatoondata.SplatoonClip;
+import tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata.SplatoonClip;
 
 import java.util.List;
