+++ src/main/java/tv/strohi/twitch/strohkoenigbot/data/repository/splatoondata/SplatoonGearRepository.java
@@ -3,8 +3,8 @@ package tv.strohi.twitch.strohkoenigbot.data.repository.splatoondata;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import tv.strohi.twitch.strohkoenigbot.data.model.splatoondata.SplatoonGear;
-import tv.strohi.twitch.strohkoenigbot.data.model.splatoondata.enums.SplatoonGearType;
+import tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata.SplatoonGear;
+import tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata.enums.SplatoonGearType;
 
 import java.util.List;
