+++ src/main/java/tv/strohi/twitch/strohkoenigbot/data/repository/splatoondata/SplatoonStageRepository.java
@@ -2,7 +2,7 @@ package tv.strohi.twitch.strohkoenigbot.data.repository.splatoondata;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import tv.strohi.twitch.strohkoenigbot.data.model.splatoondata.SplatoonStage;
+import tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata.SplatoonStage;
 
 @Repository
 public interface SplatoonStageRepository extends CrudRepository<SplatoonStage, Long> {
