+++ src/main/java/tv/strohi/twitch/strohkoenigbot/data/repository/splatoondata/SplatoonWeaponRequestRankingRepository.java
@@ -3,7 +3,7 @@ package tv.strohi.twitch.strohkoenigbot.data.repository.splatoondata;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import tv.strohi.twitch.strohkoenigbot.data.model.splatoondata.SplatoonWeaponRequestRanking;
+import tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata.SplatoonWeaponRequestRanking;
 
 import java.util.List;
