+++ src/main/java/tv/strohi/twitch/strohkoenigbot/splatoonapi/results/GearExporter.java
@@ -3,8 +3,8 @@ package tv.strohi.twitch.strohkoenigbot.splatoonapi.results;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import tv.strohi.twitch.strohkoenigbot.chatbot.spring.DiscordBot;
-import tv.strohi.twitch.strohkoenigbot.data.model.splatoondata.SplatoonGear;
-import tv.strohi.twitch.strohkoenigbot.data.model.splatoondata.enums.SplatoonGearType;
+import tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata.SplatoonGear;
+import tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata.enums.SplatoonGearType;
 import tv.strohi.twitch.strohkoenigbot.data.repository.splatoondata.SplatoonGearRepository;
 import tv.strohi.twitch.strohkoenigbot.splatoonapi.model.SplatNetGear;
 import tv.strohi.twitch.strohkoenigbot.utils.DiscordChannelDecisionMaker;
