+++ src/main/java/tv/strohi/twitch/strohkoenigbot/utils/JavaArgumentEvaluator.java
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Component;
 import tv.strohi.twitch.strohkoenigbot.StrohkoenigbotApplication;
 import tv.strohi.twitch.strohkoenigbot.chatbot.TwitchBotClient;
 import tv.strohi.twitch.strohkoenigbot.data.model.Configuration;
-import tv.strohi.twitch.strohkoenigbot.data.model.SplatoonLogin;
+import tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.SplatoonLogin;
 import tv.strohi.twitch.strohkoenigbot.data.model.TwitchAuth;
 import tv.strohi.twitch.strohkoenigbot.data.repository.ConfigurationRepository;
 import tv.strohi.twitch.strohkoenigbot.data.repository.SplatoonLoginRepository;
